[2장_이동훈] 테스트 코드를 위한 클래스 수정

+++ src/main/kotlin/hoondong/chapter02/Screening.kt
@@ -24,6 +24,6 @@ class Screening(
     }
 
     private fun calculateFee(audienceCount: Int): Money {
-        return movie.calculateMovieFee(this).times(audienceCount)
+        return movie.calculateMovieFee(this).times(audienceCount.toDouble())
     }
 }
